+++ CHANGELOG.md
@@ -14,10 +14,12 @@ The format is simplified version of [Keep a Changelog](https://keepachangelog.co
 - Add random library sort
 - Add the ability to save search queries
 - Add toggle to enable/disable hide source on swipe (@Hiirbaf)
+- Add the ability to mark duplicate read chapters as read (@AntsyLich)
 
 ### Changes
 - Temporarily disable log file
 - Categories' header now show filtered count when you search the library when you have "Show number of items" enabled (@LeeSF03)
+- Chapter progress now saved everything the page is changed
 
 ### Fixes
 - Allow users to bypass onboarding's permission step if Shizuku is installed